internal/perforce: Create package for perforce changelists
And reuse in existing places.
